refactor: 유효성 검사 로직 통합 및 DTO 구조 적용
- 각 컨트롤러에서 중복으로 처리했던 유효성 검사 로직을 GlobalException에서 공통 로직으로 통합
- 도메인 객체 대신 DTO 구조 전반적으로 적용
- DTO 적용하면서 Java Bean Validation 사용, 세부적인 예외 검증 로직도 추가

+++ src/main/java/com/readingtracker/boochive/domain/PurchaseHistory.java
@@ -9,6 +9,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -39,7 +40,7 @@ public class PurchaseHistory implements Own {
 
     private String purchaseFrom;
 
-    private Integer price;
+    private BigDecimal price;
 
     private String memo;
 
@@ -63,7 +64,7 @@ public class PurchaseHistory implements Own {
     /**
      * 구매 이력 정보 변경
      */
-    public void updateHistory(LocalDate purchaseDate, String purchaseFrom, Integer price, String memo) {
+    public void updateHistory(LocalDate purchaseDate, String purchaseFrom, BigDecimal price, String memo) {
         this.purchaseDate = purchaseDate;
         this.purchaseFrom = purchaseFrom;
         this.price = price;
